add date, service, price
impossible de tester si tout marche pour les prix, il plante quand je veux add les rooms et cars

Tous les autres changements c'est les tiens que j'ai rajouté sur ma branche

+++ src/domain/CarModel.java
@@ -5,7 +5,7 @@ import java.util.Objects;
 class CarModel {
     //Value Object
     private final String carName;
-    private final int carPrice;
+    private final float carPrice;
     CarModel(String carName, int carPrice) {
         this.carName = carName;
         this.carPrice = carPrice;
@@ -14,7 +14,7 @@ class CarModel {
     public String getCarName() {
         return carName;
     }
-    public int getCarPrice(){
+    public float getCarPrice(){
         return carPrice;
     }
